Add hook for platform codec support
Add virtual to our Skia Factory to allow clients to give us access to the Platform's native codecs.

Targeted for Android and iOS, where we'll have to do some marshaling between C++ and the host languages. However, if it works there, we can consider removing some of the codecs currently backed into our Skia build.

For power-users in the Editor, we can document that we support *all* image formats on the respective platforms, in case they want to take advantage of a specific format on a given platform (e.g. HEIF on iOS).

Viewer change:

To test the viability, viewer is modified to call into Skia's wrapper for CGImageSource. This is the Mac/iOS native image decoder. Thus viewer now has access to all of the platform's image formats.

Diffs=
300e4dc14 Add hook for platform codec support