Implement RenderFont using CoreText
Create a backend for RenderFont that uses Apple's CoreText, so we don't have to include a copy of HarfBuzz (~750K).

Had to (re)discover that Apple tries to 'help' when a font as an optical-size variation axis (opsz). In this case, Apple auto-sets the axis value to the font's pointsize... even if the caller didn't want that. Since we want to have our fonts at a huge size (to keep precision), this code has to...

1. Determine if the font has a opsz axis
2. If so, use that axis' default value for the font's pointsize, and
3.    set the font's xform to account for that (resulting in the expected large size)

Diffs=
db49cdfaa Implement RenderFont using CoreText