Use BitmapFactory for platform decoding
Extends SkiaFactory to call into Java to use Android's BitmapFactory to try to decode an image. This is called by SkiaFactory's base implementation when Skia itself fails to decode an image.

Android has a NDK api for image-decoding, and it looks like something we could use. It should be more efficient than this code, since it (1) avoids JNI up and down calls, and avoids copies of the input and output data. However, it is only available since SDK 30, and currently we support devices older than that.

Diffs=
06f0b2c28 Use BitmapFactory for runtime decoding