remove unneccessary artboard name check
get `artboardNames` turns out to create artboard instances for every artboard before getting their names. this before setting the artboard name check is just a precursor to a later error message IF an artboard with the given name could not be found.

We may not need the 2nd check (even though that won't be executed if valid artboard names are provided). We definitely do not need the 1st check.

Diffs=
1cd935cdf remove unneccessary artboard name check