Restructure to make font-fallback easier
Just pulls inner loop contents into a helper function. This is a precursor to trying to add font-fallback, where we have to examine a given output run, and look for spans of zeros (in glyphs[]), and then re-call harfbuzz on those spans with a different font(s).

Not functionality change.

Diffs=
09df6f0b2 Restructure to make font-fallback easier