Fallback for harfbuzz
Plumbing is mostly in place
Testing is just a start : have hard-coded a specific unicode font (will improve in later PRs)

Design

Harfbuzz backend has a function-pointer it will call to try to get addtional fonts to fill in the gaps (missing glyphs).
It passes a span of some of the missing characters as a clue (though in this impl, we always just return the same
fallback font).

Likely to do a better job in CJK, we may want to know some language information from the client -- which we can
pass along to the fallback-proc.

Diffs=
b4eb6b315 Fallback for harfbuzz