Don't manually flip cgimages
CGImage is flipped vertically relative to what we expect (for our texture coordinates). Previously we manually flipped the image (by drawing into a software canvas and extracting a new image). This is ok, but forces a particular colorspace and pixelformat -- rather than leaving it as the coregraphics decoder chose.

This PR removes the manual flipping, and just appends an additional transform to the CTM when we reference the image. For some tests (e.g. hero_editor) this shows a speed-up as well (not sure, but likely the colorspace of the original image is closer to what CG wanted for drawing.)

Diffs=
1f9ff75d9 Don't manually flip images