Fix a problem where we access deleted artboards and state machines
This fixes an issue around accessing data that has been deallocated.
https://2dimensions.slack.com/archives/CLLCU09T6/p1660763844645649

 Specifically, previous to this, running reset twice in a row, would crash the app, as we try to get a name of a deleted artboard instance.

Also moving artboards life cycle into the "dispose" path, as before, changing the selected artboard would dispose the old artboard (which trickles down to all the stateMachineInstances etc). Now, this will only get cleaned up once the renderer is disposed.

This is an attempt to address an issue you can run into when holding references to stateMachineInstances, in the instance where this is helpful, the stateMachine instance is simply used to grab the names of its inputs, so a ui could change the state machines input.

I *think* changing this lifecycle makes sense still, but of course this could start to have some unintended consequences, as the artboard instance referred to by the statemachine instance is no longer playing.

But in this case those old state machine instances were used to get "generic" information about the contents of the Rive object, not as concrete references to "state machine instances". I think this problem is a side effect of removing the non instanced version of artboards/state machines/animations as users now have no choice but to grab this information from instances.

perhaps a generic "hey whats in my rive file" function that offers a simple reference structure users can use will help with this though

wanna hear peoples thoughts.

- [ ] add a test to show the crash on double `reset()`

Diffs=
9036f5031 Fix a problem where we access deleted artboards and state machines