Simpify RawPath::Iter
Implement Iter as an STL-style iterator that can be traversed very
efficiently using range-for:

  for (auto [verb, pts] : rawPath) { ... }

Manually inject the implicit moveTos in RawPath (namely, the implicit
moveTo(0) at the beginning of a path, if the client didn't add one, and
a similar moveTo after close).

The implicit moveTos fix a bug in RawPath::bounds where we were not
accounting for the implicit moveTo(0), as well as a bug in
ContourMeasureIter::tryNext where we expected the first contour in a
path to begin with PathVerb::move.

With the implicit moveTos guaranteed to be in the raw path, the iterator
can be simplified to just peek back one point and give a contiguous array
beginning with p0 for each verb. In addition to generally simplifying
things, one more perk of having all the points contiguous is that it
also enables fast SIMD loads.

IterateRawPath bench result:

  MacOS NEON: 8.71ms -> 5.02 (58%)
  Windows SSE: 5.63ms -> 4.20 (75%)

BuildRawPath bench result:

  MacOS NEON: 2.09ms -> 2.19 (105%)
  Windows SSE: 2.18ms -> 2.27 (104%)

Diffs=
fdeff54d1 Simpify RawPath::Iter (#4157)