Cleaning up libpng premake to be isolated/stand-alone.
libpng_premake is now standlone!

You should just need to dofile the premake5_libpng.lua file. This will expose a variable called "libpng" which you can then use to point to include dirs, see how the viewer premake does this:
https://github.com/rive-app/rive/blob/3a8b2203e2ed1b0c57697c534415c6dede99b792/packages/runtime/viewer/build/premake5_viewer.lua#L117

You link to it by name:
https://github.com/rive-app/rive/blob/3a8b2203e2ed1b0c57697c534415c6dede99b792/packages/runtime/viewer/build/premake5_viewer.lua#L124

This should automatically setup the dependency tree for building in correct order, but I found that on some versions of I premake I need to force the order by telling the project it depends on a dependency like so:
https://github.com/rive-app/rive/blob/3a8b2203e2ed1b0c57697c534415c6dede99b792/packages/runtime/viewer/build/premake5_viewer.lua#L26

This just ensures that libpng always builds before rive_viewer.

Diffs=
404a04d35 Cleaning up libpng premake to be isolated/stand-alone. (#4713)