Fix condition with trim paths where all contours are 0 length.
Fixes a crash which would occur when trying to extract segments from a path with 0 length. This would occur when scale was set to zero (via animation or otherwise). Adds a test which recreates the condition. Prior to the fix the test would crash.

Issue discussed in Slack here: https://2dimensions.slack.com/archives/CLLCU09T6/p1674648327503309

Diffs=
42a0377bc Fix condition with trim paths where all contours are 0 length. (#4722)