Renderer lifecycle compliance
This is now managing the renderer lifecycle more robustly by instantiating the renderer directly in `RiveTextureView`. This is achieved by using an abstract factory called `createRenderer()` which needs to be overridden by subclasses.

`LowLevelActivity` in particular should show how things are much cleaner to override for a custom renderer and there's less overhead for the implementers: they're guaranteed that the lifecycle is respected within the `onAttachedToWindow() -> onDetachedFromWindow()` lifespan.

In particular, when a `View` is recycled the order of things isn't exactly guaranteed because the renderer object is deleted by an underlying async operation in a separate thread.
So, currently on 4.3.0, when scrolling quickly in the `RecyclerView` (in the emulator was more obvious than on my physical device) the example app would crash.

That's because the order of operations would be the following:
1. `onDetachedFromWindow` (queues renderer deletion)
2. `onAttachedToWindow` when recycling the view (uses the old `renderer` Kotlin object because `delete` hasn't happened yet on the background thread, but the `View` is still alive)
2.1 `renderer` resources resources are set to its new configuration and init'd (i.e. file, animations, etc.)
3. `delete` is finally called as requested from 1. and the visible resources and the renderer are removed from under the `View` that goes 💥

In order to fix this we need to guarantee that we never reuse the same Kotlin `renderer` when recycling: `renderer`s are always created `onAttachedToWindow()` and removed `onDetachedFromWindow()`. That is: If we need a new renderer, make a new instance of it altogether instead of trying to "recycle" only its Kotlin side.

Diffs=
3458576b2 Renderer lifecycle compliance (#5051)