Android - EGL & Skia Context Sharing
New version of the runtime that shares the EGL Context across multiple threads & uses a Singleton with a `shared_ptr` to auto-cleanup when the last reference to it is gone.
I also found a few random leaks peeking through the code and cleaned those up. Memory pressure has definitely gone down a lot, especially considering that we're not allocating so many EGL/Skia Contexts. I still see a lot of memory being allocated in scrolling lists and a bit of it retained after closing Activities.
Also, I've cleaned up the mechanism for stopping threads, which are now properly terminating using `join()` and cleaning up the old resources associated with their state.
In any case, this is definitely an improvement.

Future exploration:
use a single thread for rendering instead of using a single thread per View. This'd get rid of mutexes that are currently necessary to properly bind `eglContext`

Diffs=
33c5128d1 Android - EGL & Skia Context Sharing (#5250)