Use lifecycle observer to clean up resources on destroy
This addresses a problem the Duos were seeing with a memory leak where RiveAnimationVIew was instantiated, the resource/file was set but was never released because the View never became visible (i.e. `onAttachedToWindow()` never got called)
They sent over an example with this DefaultLifecycleObserver which gives us a few callbacks into the View from the surrounding Activity/Fragment. In particular, the `onDestroy()` callback helps us in this situation to clean up the resources allocated in the controller, and it also allows us to restore the File initialization inside the constructor which speeds things up a bunch.

This also fixes a null-pointer deference crash happening in our tests when we destroy `SkiaContextManager` without having allocated a `mSkContext`.

This PR also cleans up a bunch of linter warnings in our documentation within the `RiveAnimationView`.

One interesting testing exploration that could be [this AcitivityScenario](https://developer.android.com/guide/components/activities/testing), which might allow us to properly test the Activity/Fragment lifecycle without hacking together all those hooks/mocks, but I haven't gotten to it just yet.

This isn't currently breaking any of or APIs, but it might interfere with our user's APIs (namely, the Duos), but I don't think this is worth upgrading a major for -- maybe a minor though.

Diffs=
1c643fe21 Use lifecycle observer to clean up resources on destroy (#5262)