`play()` function restarts paused animations
Zach reported a problem with the current behavior:

If we call `RiveAnimationView.pause()` and then call `RiveAnimationView.play()` (the overload without parameters) the View doesn't necessarily restarts the paused animation(s) because `play()` is currently playing the first available animation in the file, which aren't necessarily the animation(s) that have been paused.

I'd like to change the API to allow play() restarting paused animations if there are any. If not, start the first available animation (timeline or state machine)

Since this is an API change, we'll need to bump a major.

Diffs=
b4d6560eb `play()` function restarts paused animations