C++ Style guide
I'm targeting only the comment from the other day and a few other simple things I could quickly get in from the style guide.
But, by looking at the Notion doc, there's quite a few more things that we'd need to adjust. Do we want to do this all in one go? I can use this PR to do all of them. Otherwise, we can get this in quickly and sort the rest of the style guide with a separate PR

Diffs=
8e5424e78 C++ Style guide (#5283)