State Machine instance doesn’t add dependencies
Unfortunately, I haven't been able to reproduce crashes in isolation.

But, we definitely don't need those object in `StateMachineInstance.dependencies()` - in fact we already have comments in the code:

https://github.com/rive-app/rive/blob/61a7070dcacb8a7eb33f603d0f28c404bfc33949/packages/runtime_android/kotlin/src/main/java/app/rive/runtime/kotlin/core/SMIInput.kt#L10-L12

and also

https://github.com/rive-app/rive/blob/61a7070dcacb8a7eb33f603d0f28c404bfc33949/packages/runtime_android/kotlin/src/main/java/app/rive/runtime/kotlin/core/LayerState.kt#L8-L10

`NativeObject.dispose()` calls `cppDelete()`, but for these two types we don't have the native deleter defined because it's already managed by the State Machine Instance.
This isn't crashing in my tests on my phones & emulators - could this lead to crashes in other implementations? Maybe? Since adding these to `dependencies` is unnecessary, it's better to remove them anyway.

Diffs=
68dd53a93 State Machine instance doesn’t add dependencies (#5350)

Co-authored-by: Umberto Sonnino <umberto@rive.app>