Runtime/interpolate states cpp
Quite a lot of churn here, I created this off of https://github.com/rive-app/rive/pull/5429 which already generated a lot of code.
On top of it, I also generated the C++ classes, which in turn added a bunch of other missing things related to text.
There's *not that much non-generated* code here, but I'll keep this draft for now..?

Diffs=
8b7587241 Runtime/interpolate states cpp (#5437)