Add fix for lifecycle observe leak
Following a leak report, `RiveAnimationView` will now initialize the lifecycle observer with a controller parameter to allow for it release upon destruction. This still allows us to behave right when a RecyclerView allocates new `RiveAnimationView`s that never get attached to the window (and they can subsequently be released when the surrounding `Activity` is closed) but also to timely release resources when `RiveAnimationView` lives within a Fragment, so it doesn't have to wait until the whole Activity is closed. More specifically, it wasn't just the native objects that were leaking but the entire View due to the `inner RiveViewLifecycleObserver`. With the class now living outside `RiveAnimationView` we get rid of that leak.
In any case, to avoid contention between `LifecycleOwner`s, when `onAttachedToWindow()` is called, the `View` will now validate the current `lifecycleOwner` and swap it out if necessary.

It also add an example Activity that was reproducing the issue mentioned above.

Diffs=
fee8fbb01 Add fix for lifecycle observe leak (#5564)

Co-authored-by: Umberto Sonnino <umberto@rive.app>