Rive Text for iOS and Android!
Adding rive text support to ios.

@mjtalbot Sorry for the shell script format churn, this is my formatter in case we want to standardize on one:
https://github.com/foxundermoon/vs-shell-format
```
  "[shellscript]": {
    "editor.defaultFormatter": "foxundermoon.shell-format"
  },
```

Also, I couldn't test the MacOSX build thoroughly as I was getting a Metal error on run which wasn't related to the text changes. But maybe I'm running the wrong project (this is the iOS one which I thought we were sharing)? Not sure which one I should run for Mac. This is the error:
<img width="847" alt="CleanShot 2023-07-20 at 16 29 56@2x" src="https://github.com/rive-app/rive/assets/454182/6c2f70f9-c8f4-4b50-a3cb-753a861e2e0b">

Diffs=
13863bf8f Rive Text for iOS and Android! (#5634)

Co-authored-by: Luigi Rosso <luigi-rosso@users.noreply.github.com>