Rive Renderer Ref
Trying to make this a bit more robust by providing a hash reference to our downstream repos in order to pinpoint which commit was used to build which Android version.

In this iteration, we're committing a new file called `.rive_renderer` to the downstream repo and when we clone, that ref is used to checkout the correct version.

Diffs=
2fbc74314 Rive Renderer Ref (#5652)

Co-authored-by: Umberto Sonnino <umberto@rive.app>