feat: add events on Android
Adds events on Android.

Topics for discussion and other considerations:
- Listeners
- URL events and us handling it automatically
- I opted to always return `properties` as a value (non nullable) and instead just return an empty map
- I'm also exposing a `data` on the `RiveEvent` which just returns all of the information as a map. Maybe not that useful but I left it in as I added that first for debugging.

Todo:

- [ ] Tests

## Listeners
Instead of adding it to the current `Listener` interface I instead created a new HashSet called `eventListeners` and expose two new methods `addEventListener` and `removeEventListener`.

For the following reasons:
- I think it'll be annoying to have to override all of the methods in the interface just to be able to use events
- Events replace some of these current handlers. `notifyStateChanged` is deprecated in Flutter and I'm assuming it should be for all of the runtimes.
- On Flutter (and similar discussion for web) we're exposing `addEventListener`/`removeEventListener` so there will be consistency.
- Our Interface approach makes deprecations difficult as you can't deprecat an overridden method (Or you can but it won't propagate down to the users. Same issue for iOS)

I'm not sure if there is a better way to structure the code considering that in its current form there is only one method in the interface: `notifyEvent`. But maybe that should be a consideration where instead of `notifyEvent` we expose:
- notifyURLEvent
- notifyGeneralEvent

The problem here is the same as above. If we ever add a new method it'll be a breaking change (unless we invoke the methods on the interface), and deprecations become difficult.

## URL Events
- Should we do this automatically for users? Or are examples enough? If we automatically open a browser when these events are triggered then should we expose a way to disable it from the animation object? You may want to disable it during debugging or some other reason.
- Is the `target` property needed on Android?

Diffs=
1f0ccae8d feat: add events on Android (#5894)

Co-authored-by: Gordon <pggordonhayes@gmail.com>
Co-authored-by: Umberto Sonnino <umberto@rive.app>