Fall back on Skia when PLS isn't supported on Android
Adds a check to make sure PLS is supported, and falls back on Skia if not.

Also keeps the PLS worker alive when the external ref count reaches zero, and instead frees its GPU resources. This has a similar effect, without having to pay the hefty price of destroying and re-creating the entire GL context

Diffs=
c4bdda67c Fall back on Skia when PLS isn't supported on Android (#6080)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>