feat: add preview build variant to android example
This adds an additional build variant to the Android example app, called `preview`:
- set to be the default (which needs to be tested, I believe this is only for the first time you open the project locally)
- this variant will use the latest hosted version of Rive Android
- additional README information to configure build variants

With this, it should be much easier for people to evaluate the example, without needing to do any additional config locally.

What this means for us is that when developing locally, ensure you have the correct build variant set: `debug`.

This PR also replaces the network image that is loaded in the first example page, that has been causing issues. Requires further investigation but keeping it in is a bad experience for anyone running the app themselves.

Diffs=
0c4483cd0 feat: add preview build variant to android example (#6088)

Co-authored-by: Gordon <pggordonhayes@gmail.com>