Start synchronizing with a File lock
First stab at hunting a few race conditions, I've put together some examples that try to expose race conditions but I've only found a scenario on `scheduleFrame()` for now.
In any case, I've resorted to using a `ReentrantLock` on a `File` object, since everything originates from there, and trying to gate the most probable sources of resource contention, including `disposeDependencies()` on the `RiveArtboardRenderer`.

I still need to take a look at the input queue. And, I'd like to remove some of the old `RiveArtboardRenderer` APIs that have been moved to the `RiveFileController`, some of which are already deprecated
With this considerations, I'm pretty sure this will need a major bump.

Diffs=
c4a8015a8 `synchronize{}` with a ReentrantLock (#6170)

Co-authored-by: Umberto Sonnino <umberto@rive.app>