Queue Inputs
First implementation on the input queue to avoid locking on the whole file when processing inputs.

We currently synchronize the input operations on the queue itself, since queuing elements and processing them happen on two separate threads.
Since we're using a [thread-safe queue](https://developer.android.com/reference/java/util/concurrent/ConcurrentLinkedQueue), we could also be doing something like this and avoid locking altogether:
```kotlin
private fun queueInput(stateMachineName: String, inputName: String, value: Any? = null) {
    changedInputs.add(ChangedInput(stateMachineName, inputName, value))
    // Restart the thread if needed.
    onStart?.invoke()
}

@WorkerThread
private fun processAllInputs() {
    var input = changedInputs.poll()
    while (input != null) {
        // Input fire/set & play()
    }
    input = changedInputs.poll()
}
```

Curious to hear your thoughts @csmartdalton and @dragostis!

Still to-do:
- [x] Remove deprecated thread-unsafe methods from `RiveArtboardRenderer`

EDIT:
Removed a lot of the old stuff that was still on the `RiveArtboardRenderer` but was redundant since everything was moved onto `RiveFileController`. This simplified a lot of flows and removes a bunch of deprecated methods. It's also going to require a major bump because we're removing a few APIs. But the major bump was already inevitable since we're changing the way inputs are behaving.

Diffs=
474aff99a Inputs Queue (#6177)

Co-authored-by: Umberto Sonnino <umberto@rive.app>