Setting multiple inputs at once
Fixes a bug reported by the Duos on Slack with the way inputs are processed and reduces calls to `play()` when processing multiple inputs on the same State Machine Instance.

The main problem was here: `play(stateMachineInstance, settleStateMachineState = true)` with `settleStateMachineState = true` was causing the State Machine Instance to advance immediately after setting the input, resolving the state before it's needed. This PR restores behavior pre-9.0.0 and also restores `RiveEventTest` needing to `fireState` twice, which is a different issue, but also is consistent with the other runtimes.

It also adds an API for setting multiple inputs at once & a test.

Diffs=
c0e6b78d1 Setting multiple inputs at once (#6221)

Co-authored-by: Umberto Sonnino <umberto@rive.app>