Replace intermediate flushes with "logical" flushes
Since textures have hard size limits, we can't always fit an entire frame into one flush. It's rare for us to require more than one flush in a frame, but for the times that we do, this PR adds a nested "LogicalFlush" class that can be built up into a list and executed at the end of a frame.

Diffs=