More browserstack
Ok, getting to a point where i think its worth talking it through and looking to merge / adjust.

Problems:
1) lfs files
we have device goldens (one set of goldens for each test per device)
and generic goldens (one set of goldens for each test)

Those device goldens are a problem, it'll increase the checkout time a bit for fresh checkouts. its 300MB worth of images,  its in LFS I guess. But still feels like a lot?could do 720p or even 540p (would probably need to do the same for the more general golden target though, otherwise we are creating two which will suck) or could just not do device checks right now, and soley rely on generic golden checks.

we can however limit how you check out the repo for github actions if we don't want other actions to to check out these goldens, so that shouldn't be a problem...

2) organisational
should probably decide where a few folders go, i just "stuck em" somewhere convenient. but i think there's probably a better place...

3) browserstack tests take too long.
right now the browserstack tests take a long time, and I think we want to get to a place where they take less time. A big help here would be to move on from the checks once the first frame is rendered and written to disk automatically. and at the end go to a "completed" page. that way our browserstack test just needs to spin up the app and check which page we're on, once we are completed, we can exit.

4) golden management
its something I've not spent any time doing. But I think as a minimum we want a script to update the goldens based on some candidates (locally or remotely). I think this can come later though as we currently do not have test failures block developement.
later it would be great to be able to do this from the comparison pages

Diffs=
9ce998416 More browserstack (#6446)

Co-authored-by: Maxwell Talbot <talbot.maxwell@gmail.com>