Implement clipping with stencil in MSAA mode
This approach is slightly unique because we actually have to clear clips out of the stencil buffer when we're done with them. To accomplish this, we add a new type of PLSDraw to our system that clears or intersects the stencil buffer. These stencilReset operations can be re-ordered and batched together when non-overlapping regions have different clips.

Also adds a system to track clip "contentBounds" and "readBounds". For this change we ended up only needing the contentBounds, but tracking readBounds as well lays the groundwork for incorporating scissor optimizations into clipping.

To keep things simple, we don't reverse-sort clipped opaque paths in this change. We can work out the details on that later.

Diffs=