Implement atomic mode for GL framebuffer fetch
Qualcomm has a "noncoherent" framebuffer fetch extension that allows us to store PLS in tiled memory, but without the cost of having to enable raster ordering. Since atomic mode is already implemented, we can just use that mode in this case.

Since the coverage plane requires actual atomic operations, we leave that plane in a storage texture. Hopefully the tiler does something smart with the hint that we only use glMemoryBarrierByRegion.

Diffs=