Specialize SPIR-V for pls::ShaderFeatures
Add specialization constants to PLS shaders when compiling to SPIR-V. Set these constants on VkPipelines based on pls::ShaderFeatures.

Add a new shader switch, "@FIXED_FUNCTION_COLOR_BLEND", to explicitly turn on direct-to-framebuffer rendering. We used to implicitly make this switch based on @ENABLE_ADVANCED_BLEND, which broke the model that a superset of shader features was always compatible with any subset. This also allows us to specialize Vulkan shaders before making the switch to direct-to-framebuffer rendering.

Diffs=