Add better stack trace & test
Record the StackTrace for when NativeObjects are released and tag it along when trying to access an invalid pointer.
This should help Duos give us more insights into what's happening that's causing crashes for them.

The primary suspect is that some combination of API calls is causing them to release C++ objects unintentionally, but until we get a repro or some actual code from them, we can't know for sure.

Diffs=
f41ea82aa Add better stack trace & test (#7775)

Co-authored-by: Umberto Sonnino <umberto@rive.app>