Fallback Fonts on Android
PR is a bit chunky, but mostly because I added a few XML files for testing which are bumping up the line count.

Currently, the API looks like this:
```kotlin
fun setFallbackFont(byteArray: ByteArray): Boolean
fun setFallbackFont(opts: Fonts.FontOpts? = null): Boolean
```
Which allows our users to set fallback options with either a pre-loaded font (via bytes), or by specifying the Font Options. If none are specified, it goes back to the default
```kotlin
val DEFAULT = FontOpts(
  familyName = "sans-serif",
  weight = 400,
  style = "normal"
)
```
Which is the default configuration for Roboto Normal.
Our fallback font getter in C++ will also prioritize any font set by the user first and it'll try with Roboto if it can't find anything the user has set.
We could try all available system fonts to find a good match, but that could become really slow, especially considering that the callback is called every time it can't find a glyph.

I added a sample Activity to demonstrate our fallback font behavior with a riv file which looks like this:
<img width="310" alt="image" src="https://github.com/user-attachments/assets/f73be722-e08e-4e91-b3c0-3eb1a5b106a3">

There are also a bunch of tests to validate behaviors in Kotlin and in JNI/C++.

Leftovers:
- [x] Parse remaining XML files for other fallback fonts possibilities
- [x] Handle more precise font matching via `FontOpts`
- [x] Aliases need to handle the weight
- [x] Tie this together with the NDK
- ~~Lang as enum (optional)~~ Didn't find a good pattern here, I'll leave it as is because it gives us more flexibility

Diffs=
695007fd0 Fallback Fonts on Android (#7685)

Co-authored-by: Umberto Sonnino <umberto@rive.app>