fix: nested artboards advancing
Attempt nr 2.

Resolves: https://github.com/rive-app/rive-android/issues/338

With this change we will only settle if nested artboards (state machines) are also settled.

Two tests "broke" as a result of this change. But with this change, these tests behave the same as our other runtimes, meaning it is the expected behavior. This PR went through a few iterations, and this approach seems to be the best trade-off between the desired implementation and nr. of "failing" tests.

Diffs=
c65e93fee2 fix: nested artboards advancing (#8295)

Co-authored-by: Gordon <pggordonhayes@gmail.com>