editor and runtime: allow vertices in a vector nslicer to be out of bounds
Followup on #8459, where I returned a vertex untransformed if it's out of bounds of a vector n-slicer. I realized that instead of not transforming, they should be still be transformed, just not scaled. This PR translates the value by offset, and updates both the editor and the runtime. I also added a golden test for vector n-slicing.

All of these clipping changes was to support this leaf animating. Note that a vertex temporarily goes out of bounds:
![vec_oob](https://github.com/user-attachments/assets/5d3eaacf-4711-48cb-8318-a30b152a1864)

About the golden test:
![image](https://github.com/user-attachments/assets/c4dad5b7-4b4f-4de7-9150-156c2fac828c)

Top to bottom,
1. base case, the two 100x100 rectangles should be spaced out
2. same as 1, except the right rectangle should have a radial gradient with a yellow 'halo' in the center
3. the two shapes should look like complete stars. Before this PR btw, the stars would either be cut off or one of the vertices would 'fold back':
![image](https://github.com/user-attachments/assets/305936e8-80bf-4d13-9fe8-43b7d35fbf9f)

4. the 'dog' should be facing right, and the tail should have 4 dashes on it. The dog is actually 'flipped' with a negative width. It also contains a group of vectors. So this test case is making sure that a recursive structure works, as well as negative width.

Diffs=
33dc66a73e editor and runtime: allow vertices in a vector nslicer to be out of bounds (#8495)