Add AdvanceFlags
Initially this PR was to fix a bug preventing LayoutComponents from animating when the state machine isn't running (and also in design mode), unlike NestedArtboards.

In the process, we decided to convert advance params to flags to allow less effort if its necessary to add more params in the future.

Diffs=
17cbda4b69 Add AdvanceFlags (#8492)