JNI check and rethrow
This comes from a Duo [request on Slack](https://2dimensions.slack.com/archives/C07BR8N0XMK/p1728678603588259) and is a first step towards getting more context for crashes that happen on our render thread but the exception is actually thrown in Java

It also bumps the SDK to the latest.

**EDIT:**
Here's a screenshot for this handler in LogCat
![image](https://github.com/user-attachments/assets/9f294623-a23b-4deb-adfe-3376e2562528)

This is just by adding the following in `RiveFileController @ advance()`
`throw RiveException("This is a test exception for JNI")`

Before this, the runtime would ignore the error until another JVM -> JNI call was made, which is why they'd be reporting crashes on e.g. `FindClass("app/rive/runtime/kotlin/core/Fit")`

Diffs=
50e31c53f0 JNI check and rethrow (#8497)

Co-authored-by: Umberto Sonnino <umberto@rive.app>