Add GitHub action for re-releasing
Now we can attempt to re-release to Maven Central without changing the version number.

Moves steps into composite actions which are referenced by the script. I went with that approach for now since it's the most 1:1 with what we have, but I think in the future we should investigate reusable workflows. I'm worried about log visibility with this approach.

I'm also not sure if the pathing was correct when referencing the actions. I believe it's from the repository root but we'll see.

If this is successful, we can consider moving `release.yml` next.

GitHub articles of use:
[Comparison between composite actions and reusable workflows](https://docs.github.com/en/actions/sharing-automations/avoiding-duplication)
[Reusable workflows](https://docs.github.com/en/actions/sharing-automations/reusing-workflows)
[Composite actions](https://docs.github.com/en/actions/sharing-automations/creating-actions/creating-a-composite-action)

Diffs=
bedf8844a4 Add GitHub action for re-releasing (#8627)

Co-authored-by: Erik <erik@rive.app>