fix dat converter formula crash
the actual crash was caused by using size_t instead of int for the index.
Since it would be set to -1 at the last loop, being a size_t would set it to max int and overflow causing a crash.
But for safety I decided to change when converters are deleted, following how file assets are stored and only deleted when the file is deleted.

Diffs=
df6ccaa5ff fix dat converter formula crash (#9168)