fix: Remove printmapping ProGuard rule
# Description
This was forcing a particular name to the ProGuard mapping output. If clients want mapping, they can add it to their own ProGuard rules.

# Details
[Context in Slack](https://2dimensions.slack.com/archives/C073UBSHFB7/p1739547229104789).

Technically this is a breaking change, as the rules in `consumer-rules.pro` are merged into consuming apps. However, they can always add it back if they need it, with the advantage of getting to name it whatever they want. It's not the job of libraries to force this.

Diffs=
3697d133a0 fix: Remove printmapping ProGuard rule (#9191)

Co-authored-by: Erik <erik@rive.app>