fix(Android): Render thread crash
# Description
The render thread was not synchronizing during initial drawing, meaning if a user changed the File during this time, there would be a crash after the last reference to the File was released.

# Details
[Slack context (Duolingo)](https://2dimensions.slack.com/archives/C029X99PETE/p1743109091460179)

Occurs when the main thread is does a `release()` of the Controller's File, because the user has called `setRiveFile` to replace it with a new one. Meanwhile, the worker thread is mid-render, and calling `Artboard::draw`, in turn calling `cppDrawAligned(cppPointer, ...)`. `cppPointer` having a getter, checking that it's not null, which at this point it is.

We have `synchronized(lock)` around both the `File::release()` implementation and the `Artboard::draw`, but that just ensures they're not running at the same time. If `release` gets called first, `draw` will block until it finishes, and once finished it's internal C++ pointer has been set to null.

The solution I chose was to wrap `RiveArtboardRenderer::draw` in `synchronized(controller.file?.lock ?: this) {...}`, which it's not currently.

Diffs=
5a6b6d69a2 fix(Android): Render thread crash (#9392)

Co-authored-by: Erik <erik@rive.app>