Fix race condition in Metal background shader compiler (#10355) 92de3f0a1a * Fix race condition in Metal background shader compiler
Setting an atomic variable outside of a std::condition_variable's corresponding mutex and then notifying the condition_variable has a subtle race condition, and the ordering of operations looks like this:

Thread A: acquire mutex
Thread A: check m_shouldQuit // it's false
Thread B: m_shouldQuit = true
Thread B: m_workAddedCondition.notify_all() // nothing is currently waiting on it so this does nothing
Thread A: m_workAddedCondition.wait() // now it's waiting, having just missed the above notification

Now thread A will wait forever because no more notifies to the condition_variable are coming and, in the case of this code, the call to m_compilerThread.join() will also wait forever (since the thread will never quit), so shutting down the renderer will never finish.

Instead, m_shouldQuit should only be set to true while in the mutex (because the check of its value and the subsequent .wait call also happen within the mutex, so this prevents them interleaving). This change fixes this.

* Minor formatting adjustment to make the github workflow tests re-run