Fix the goldens by adding a missing define to the android runtime (#10525) 9d70fea69b The nightly run of android compatibilitytest was failing for the rive_mesh golden due to a mismatch between defines (specifically, the existence of RIVE_ANDROID in core but not android runtime). This change adds the known-missing define to the android build so that they match.
It would probably be a good idea to find a way to keep the defines synced between the two runtimes because it's not clear which other defines might be missing, and if any are added later to the core runtime there would be a new potential mismatch.

I have no idea what that would look like, though.

Co-authored-by: Josh Jersild <joshua@rive.app>