feat: Initial implementation of a clockwise rendering mode (#10897) 2fb75868b5 Add a new InterlockMode that overwrites all fill rules as clockwise and implements the clockwise path rendering algorithm using raster ordered PLS. The only backend to support this so far is GL with shader images, but more will come.
Notably, when there is no advanced blend, we can set
"fixedFunctionColorOutput" and render directly to the target
framebuffer.

Performance so far looks promising, especially on Intel, but more
in-depth perf optimizations arebyet to come.

This is implemented as its own InterlockMode only to give it soak time.
Once this rendering mode is stable, we can merge it back into
"rasterOrdering" and just select draw shaders based on fill rule.