fix(editor): Display runtime errors when Script input VM properties not set (#11177) cd70e30236 Fixes a crash when trying to use pointer events when a ViewModel property input is not setup properly on a scripted drawable instance.
The code was checking for ScriptAsset.vm which will always be non-null when bytecode is compiled in the editor. We should be checking instead for ScriptedObject.state because that will only be non-null if the ScriptedObject has validated (and we early out if that is null so we don't crash).

Also added some errors when VM properties are not setup correctly as Script inputs so users will be aware that there is an issue with their VM setup or bindings