feat(Android): Compose Layout and Sample App Improvements (#11249) e17d0603d9 Adds proper support for layouts in Compose. The missing pieces were resizing the artboard and adding a scale factor (though the latter is less critically important).
Also contains a number of improvements to the sample app, specifically around edge-to-edge support. When we went to SDK 36, edge-to-edge was enabled by default, but we didn't update the samples to account for them.

Adds the layout demo from the demos page to put Android in line with other SDKs.

Co-authored-by: Erik <erik@rive.app>